Add an additional correlation function from CT-L/Geostat inversions.
CarbonTracker-Lagrange and the geostatistical inversions from Anna
Michalak's group tend to use an eight-part daily cycle per week.
There are a few difficulties in translating it from that time-based
function to my time-difference-based function.  I decided on averaging
the effective correlation functions for hours equivalent to zero, one,
and two modulo three, which gave this triangular pulse.  When
translating this to the annual slots, this is equivalent to about
six-week windows, slightly longer than the month-long windows used in
TRANSCOM-era inversions.